Session robustness + session notification
Added the ability to completely disconnect from the server and maintain session state - done through AJAX requests to keep the server up to date with the current session state.
Old, invalid sessions are automatically removed (but should be impossible to create in the first place)
Notification on other pages if there is a session currently running.
A few fixes here and there